Restructure external content module, add license validator
