+++ teachbooks/external_content/__init__.py
@@ -1,0 +1,3 @@
+
+GIT_PATH = "_git"
+
